+++ src/pools/selection-strategies/fair-share-worker-choice-strategy.ts
@@ -17,7 +17,7 @@ import type {
  * @typeParam Response - Type of execution response. This can only be structured-cloneable data.
  */
 export class FairShareWorkerChoiceStrategy<
-  Worker extends IWorker<Data>,
+  Worker extends IWorker,
   Data = unknown,
   Response = unknown,
 > extends AbstractWorkerChoiceStrategy<Worker, Data, Response>