+++ src/pools/selection-strategies/least-busy-worker-choice-strategy.ts
@@ -16,7 +16,7 @@ import type {
  * @typeParam Response - Type of execution response. This can only be structured-cloneable data.
  */
 export class LeastBusyWorkerChoiceStrategy<
-  Worker extends IWorker<Data>,
+  Worker extends IWorker,
   Data = unknown,
   Response = unknown,
 > extends AbstractWorkerChoiceStrategy<Worker, Data, Response>